feat(headers): Implement Content-Language header field
Closes #475
