+++ src/header/common/mod.rs
@@ -22,6 +22,7 @@ pub use self::cache_control::{CacheControl, CacheDirective};
 pub use self::connection::{Connection, ConnectionOption};
 pub use self::content_length::ContentLength;
 pub use self::content_encoding::ContentEncoding;
+pub use self::content_language::ContentLanguage;
 pub use self::content_type::ContentType;
 pub use self::cookie::Cookie;
 pub use self::date::Date;
@@ -303,6 +304,7 @@ mod cache_control;
 mod cookie;
 mod connection;
 mod content_encoding;
+mod content_language;
 mod content_length;
 mod content_type;
 mod date;
