+++ src/header/mod.rs
@@ -19,7 +19,7 @@ use unicase::UniCase;
 use self::internals::Item;
 use error::HttpResult;
 
-pub use self::shared::{Charset, Encoding, EntityTag, HttpDate, Quality, QualityItem, qitem, q};
+pub use self::shared::*;
 pub use self::common::*;
 
 mod common;
