+++ src/header/shared/mod.rs
@@ -2,10 +2,12 @@ pub use self::charset::Charset;
 pub use self::encoding::Encoding;
 pub use self::entity::EntityTag;
 pub use self::httpdate::HttpDate;
+pub use self::language::Language;
 pub use self::quality_item::{Quality, QualityItem, qitem, q};
 
 mod charset;
 mod encoding;
 mod entity;
 mod httpdate;
+mod language;
 mod quality_item;
